requirements.txt
pillow>=9.0.0
exifread>=2.3.2
geopy>=2.3.0
numpy>=1.22.0
tensorflow-cpu>=2.10.0  # Use CPU version which is easier to install
python-dotenv>=0.21.0
pydantic>=1.10.0
python-multipart>=0.0.5
fastapi>=0.85.0
uvicorn>=0.19.0
pytest>=7.2.0